test: fold client log batch coverage into existing suite
